requirements-docker.txt
Django>=4.2,<5.1
djangorestframework>=3.14
djangorestframework-simplejwt
rest_framework_simplejwt
channels>=4.0
channels-redis>=4.1
torch==2.10.0+cpu
torchvision==0.25.0+cpu
torchaudio==2.10.0+cpu
torchcodec==0.10.0
faster-whisper>=1.2.0
silero-vad
onnxruntime>=1.14,<2
pydub>=0.25.1
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
groq
psycopg2-binary
transformers==4.33.0
numpy==1.26.4
pika
soundfile==0.13.1
django-cors-headers==4.9.0
uvicorn[standard]==0.34.0
scipy>=1.11
aio_pika==9.5.5
grpcio
grpcio-tools
TTS==0.22.0
celery